backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
websockets>=12.0
langchain>=0.2
langchain-google-genai>=1.0
google-generativeai>=0.5
networkx>=3.3
numpy>=1.26
aiosqlite>=0.20
python-dotenv>=1.0
uagents>=0.14
mcp>=1.0
